+++ lib/serialize/content.js
@@ -1,7 +1,7 @@
 
 // return summary and body
 
-var tomark = require('./tomark');
+var marked = require('./marked');
 
 module.exports = function(body) {
 
@@ -16,8 +16,8 @@ module.exports = function(body) {
     }
 
     return {
-        summary: tomark(summary),
-        body: tomark(body)
+        summary: marked(summary),
+        body: marked(body)
     }
 
 }
