+++ lib/serialize/marked.js
