+++ lib/serialize/page.js
@@ -1,7 +1,7 @@
 
 // return pages and issues
 
-var tomark = require('./tomark');
+var marked = require('./marked');
 var config = require('../config');
 
 require('./format')()
@@ -22,7 +22,7 @@ module.exports = function(issues) {
             page.title = issue.title.split(']')[1];
             page.created = issue.created_at;
             page.updated = issue.updated_at;
-            page.body = tomark(issue.body);
+            page.body = marked(issue.body);
 
             // pages push this item
             pages.push(page)
